# APIdoc.yaml
openapi: 3.0.3
info:
  title: DivingRegistrationAPI
  description: API to DivingRegistration application
  version: 1.0.0
servers:
  - url: 'https://dev-restandroid.users.info.unicaen.fr/api'
  - url: 'http://127.0.0.1:8000/api'
paths:
  /adherents:
    get:
      description: get all active divers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Adherent'
    post:
      description: insert a new active diver
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdherentWithConfirmedPassNoId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Adherent'
        '422':
          description: texts are too long or (nom,prenom) is not unique, or email is not unique, or pass is not strong enough, or not confirmed, or aptitude level is incorrect
    put:
      description: update a diver
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/AdherentWithConfirmedPass'
                - properties:
                    old_pass:
                      type: string
                      format: password
                      minLength: 8
                      example: 'MotDePasse-Pas-Securis3'
              required:
                - old_pass
                - id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Adherent'
        '404':
          description: not found
        '422':
          description: texts are too long or (nom,prenom) is not unique, or email is not unique, or pass is not strong enough, or not confirmed, or old_pass does not match current password, or aptitude level is incorrect
...
components:
...
  schemas:
    Actif:
      type: object
      properties:
        actif:
          type: boolean
          description: true iff it can be chosen for new dives
          example: true
    Adherent:
      allOf:
        - $ref: '#/components/schemas/Personne'
        - $ref: '#/components/schemas/AdherentBase'
        - $ref: '#/components/schemas/Actif'
    AdherentBase:
      type: object
      description: A diver registered in the association
      properties:
        license:
          type: string
          description: the diver's license
          example: '9242391985562582'
        date_certificat_medical:
          type: string
          format: date
          example: '2023-05-31'
        forfait:
          type: string
          description: the diver's remaining dives
          example: '75'
        niveau:
          type: integer
          description: the aptitude ID of the diver
          example: 9
...
    AdherentWithConfirmedPassNoId:
      allOf:
        - $ref: '#/components/schemas/AdherentBase'
        - $ref: '#/components/schemas/PersonneWithConfirmedPassNoId'
    AdherentWithConfirmedPass:
      allOf:
        - $ref: '#/components/schemas/AdherentWithConfirmedPassNoId'
        - $ref: '#/components/schemas/Id'
...
    Id:
      type: object
      properties:
        id:
          type: integer
          description: its unique ID
          example: 3
...
    PersonneBase:
      type: object
      description: A person registered to the application
      properties:
        nom:
          type: string
          description: the person's last name
          example: 'Dupont'
          maxLength: 45
        prenom:
          type: string
          description: the person's first name
          example: 'Dupond'
          maxLength: 45
        email:
          type: string
          format: email
          description: the person's mail address
          example: 'dupondt@iut.fr'
          maxLength: 100
    PersonneRights:
      type: object
      properties:
        directeur_de_section:
          type: boolean
          description: true iff the person is the technical director
          example: false
        secretaire:
          type: boolean
          description: true iff the person is the secretary
          example: false
        securite_de_surface:
          type: boolean
          description: true iff the person can be surface security
          example: false
        pilote:
          type: boolean
          description: true iff the person can be pilot
          example: false
    Personne:
      allOf:
        - $ref: '#/components/schemas/Id'
        - $ref: '#/components/schemas/PersonneBase'
        - $ref: '#/components/schemas/Actif'
...
    PersonneWithConfirmedPassNoId:
      allOf:
        - $ref: '#/components/schemas/PersonneBase'
        - $ref: '#/components/schemas/PersonneRights'
        - type: object
          properties:
            pass:
              type: string
              format: password
              description: the person's password
              example: 'Dtunpo-2024'
              maxLength: 60
              minLength: 8
            pass_confirmation:
              type: string
              format: password
              description: the person's password
              example: 'Dtunpo-2024'
              maxLength: 60
              minLength: 8
...